+++ lib/alga_view/alga_view.dart
@@ -24,9 +24,8 @@ class _AlgaShellRouteViewState extends State<AlgaShellRouteView> {
             const VerticalDivider(width: 1),
             Expanded(
               child: Scaffold(
-                  body: ClipRect(
-                child: widget.child,
-              )),
+                body: ClipRect(child: widget.child),
+              ),
             ),
           ],
         ),